Changed KVO for progressView to the new API and trying to replace URLSessionTask creating with the new generic type
